Removed capturing Azure credentials
Previous version of the deployment script would capture global admin AAD credentials and then store it as a processed credential object in PowerShell for limiting user logins. 

This exposes the AAD global admin login in cleartext (with its password)--which presents a greater risk for having the account compromised by bad actors. 

Utilizing the MSOnline and AzureRM login cmdlets, the script now calls forth 2 Azure logins that pass authentication into the PowerShell session, ensuring the AAD global admin credentials are never exposed in the process (in the console, in cleartext).